Standardizes error codes for all commands
